refactor(www): move data-fetching helpers into separate file
